// lib/types/character.ts
export type CharacterType =
    | 'luffy'
    | 'zoro'
    | 'nami'
    | 'usopp'
    | 'sanji'
    | 'chopper'
    | 'robin'
    | 'franky'
    | 'brook'
    | 'jimbei'
    | 'ace'
    | 'sabo'
    | 'law'
    | 'shanks'
    | 'whitebeard'
    | 'blackbeard'
    | 'kaido'
    | 'bigmom'
    | 'yamato'
    | 'dragon'
    | 'garp'
    | 'koby'
    | 'smoker'
    | 'tashigi'
    | 'buggy'
    | 'mihawk'
    | 'hancock'
    | 'crocodile'
    | 'doflamingo'
    | 'kuma'
    | 'moria'
    | 'jinbe'
    | 'teach'
    | 'akainu'
    | 'kizaru'
    | 'aokiji'
    | 'fujitora'
    | 'ryokugyu'
    | 'sengoku'
    | 'roger'
    | 'rayleigh'
    | 'oden'
    | 'rocks'
    | 'joyboy'
    | 'imu'
    | 'captain'
    | 'trafalgar_law';